requirements.txt
arabic-reshaper==3.0.0
asgiref==3.9.1
charset-normalizer==3.4.2
distlib==0.3.9
Django==5.2.4
filelock==3.18.0
pillow==11.3.0
platformdirs==4.3.8
python-bidi==0.6.6
reportlab==4.4.3
sqlparse==0.5.3
tzdata==2025.2
virtualenv==20.31.2
